Add resource for getting languages
This is one of the fields required by invoices. Now it returns just a
list of objects each of which contain a field for language code.

Maybe this should go to the database level. Then invoices could refer to
languages by id instead of maintaining a code string. We'll see.